desafio node finalizado

+++ index.js
@@ -6,7 +6,32 @@ api.use(express.json())
 
 const orders = []
 
-api.post('/order', (request, response) => {
+const checkUserId = (request, response, next) => {
+        const {id} = request.params
+        const index = orders.findIndex(ordered => ordered.id === id)
+
+        if(index < 0) {
+            return response.status(404).json({error: "Order not found"})
+        }
+
+        request.userIndex = index
+        request.userId = id
+
+        next()
+
+}
+
+const methodAndURL = (request, response, next) => {
+    const method = request.method
+    const URL = request.url 
+
+    console.log(`O método dessa requisão é ${method} e a URL é ${URL}`)
+
+    next()
+
+}
+
+api.post('/order', methodAndURL, (request, response) => {
     const {order, clientName, price} = request.body
 
     const newOrder = {id: uuid.v4(), order, clientName, price, status: 'Em preparação'}
@@ -16,35 +41,25 @@ api.post('/order', (request, response) => {
     return response.status(201).json(newOrder)
 })
 
-api.get('/order', (request, response) => {
+api.get('/order', methodAndURL, (request, response) => {
     return response.json(orders)
 })
 
-api.put('/order/:id', (request, response) => {
-    const {id} = request.params
+api.put('/order/:id', checkUserId, methodAndURL, (request, response) => {
     const { order, clientName, price} = request.body
+    const index = request.userIndex
+    const id = request.userId
 
     const updadeOrder = {id, order, clientName, price}
-
-    const index = orders.findIndex(ordered => ordered.id === id)
-
-    if(index < 0) {
-        return response.status(404).json({message: "Order not found"})
-    }
 
     orders[index] = updadeOrder
 
     return response.json(updadeOrder)
 })
 
-api.delete('/order/:id', (request, response) => {
-    const {id} = request.params
-
-    const index = orders.findIndex(ordered => ordered.id === id)
-
-     if(index < 0) {
-        return response.status(404).json({message: "Order not found"})
-    }
+api.delete('/order/:id', checkUserId, methodAndURL, (request, response) => {
+    const id = request.userId 
+    const index = request.userIndex
 
     orders.splice(index, 1)
 
@@ -52,44 +67,34 @@ api.delete('/order/:id', (request, response) => {
 
 })
 
-api.get('/order/:id', (request, response) => {
-    const {id} = request.params
+api.get('/order/:id', checkUserId, methodAndURL, (request, response) => {
+    const id = request.userId
 
     const orderFound = orders.find(ordered => ordered.id === id)
 
     if(orderFound < 0) {
-        return response.status(404).json({message: "Order not found"})
+        return response.status(404).json({error: "Order not found"})
     }
 
     return response.json(orderFound)
 
 })
 
-api.patch('/order/:id', (request, response) => {
-    const {id} = request.params
+api.patch('/order/:id', checkUserId, methodAndURL, (request, response) => {
     const { order, clientName, price} = request.body
-    const index = orders.findIndex(ordered => ordered.id === id)
+    const index = request.userIndex
+    const id = request.userId
+ 
 
     const updateOrder = {id, order, clientName, price, status: "Pronto"}
 
     orders[index] = updateOrder
 
-    if(index < 0) {
-        return response.status(404).json({message: "Order not found"})
-    }
 
     return response.json(updateOrder)
 
-
-
-    
-
-
-
-
 })
 
-
-api.listen(port, (request, response) => {
+api.listen(port, () => {
     console.log(`🚀 Servidor rodando com sucesso na porta ${port}`)
 })
